# budget-maintenance/application/server/schema/api-docs.json
{"swagger":"2.0","info":{"description":"This page lists all the rest apis for Hotel Management App.","version":"1.0-SNAPSHOT","title":"Budget Rest APIs"},"host":"localhost:8081","basePath":"/","tags":[{"name":"budget-controller","description":"Test Swagger"}],"paths":{"/add":{"post":{"tags":["budget-controller"],"summary":"Einzahlung/Auszahlung","operationId":"addPaymentUsingPOST","consumes":["application/json"],"produces":["*/*","application/json"],"parameters":[{"in":"body","name":"payment","description":"payment","required":true,"schema":{"$ref":"#/definitions/Payment"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Payment"}},"201":{"description":"Created"},"400":{"description":"Ungültige Anfrage"},"401":{"description":"Unauthorized"},"403":{"description":"Unerlaubter Zugriff"},"404":{"description":"Not Found"},"500":{"description":"Ein allgemeiner Serverfehler ist aufgetreten"}},"deprecated":false}},"/delete/{year}/{month}":{"delete":{"tags":["budget-controller"],"summary":"Monatsliste löschen","operationId":"deleteUsingDELETE","produces":["*/*","application/json"],"parameters":[{"name":"month","in":"path","description":"month","required":true,"type":"string"},{"name":"year","in":"path","description":"year","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/MonthlyModel"}},"204":{"description":"No Content"},"400":{"description":"Ungültige Anfrage"},"401":{"description":"Unauthorized"},"403":{"description":"Unerlaubter Zugriff"},"500":{"description":"Ein allgemeiner Serverfehler ist aufgetreten"}},"deprecated":false}},"/stat":{"get":{"tags":["budget-controller"],"summary":"Statusabfrage","operationId":"askForStatUsingGET","produces":["*/*","application/json"],"parameters":[{"name":"month","in":"query","description":"month","required":true,"type":"integer","format":"int32"},{"name":"year","in":"query","description":"year","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/MonthlyModel"}},"400":{"description":"Ungültige Anfrage"},"401":{"description":"Unauthorized"},"403":{"description":"Unerlaubter Zugriff"},"404":{"description":"Not Found"},"500":{"description":"Ein allgemeiner Serverfehler ist aufgetreten"}},"deprecated":false}}},"definitions":{"MonthlyModel":{"type":"object","properties":{"auszahlungen":{"type":"array","items":{"$ref":"#/definitions/ReturnPaymentOut"}},"einzahlungen":{"type":"array","items":{"$ref":"#/definitions/ReturnPaymentIn"}},"kassenbestand":{"type":"number"},"name":{"type":"string"}},"title":"MonthlyModel"},"Payment":{"type":"object","properties":{"betrag":{"type":"number"},"kategorie":{"type":"string"},"memo":{"type":"string"},"typ":{"type":"string","enum":["EINZAHLUNG","AUSZAHLUNG"]}},"title":"Payment"},"ReturnPaymentIn":{"type":"object","properties":{"betrag":{"type":"number"},"datum":{"type":"string"}},"title":"ReturnPaymentIn"},"ReturnPaymentOut":{"type":"object","properties":{"kategorie":{"type":"string"},"summe":{"type":"number"}},"title":"ReturnPaymentOut"}}}