# budget-maintenance/application/server/src/main/java/com/uhe/app/server/api/StatApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.2.3).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.uhe.app.server.api;

import com.uhe.app.server.model.MonthlyModel;

import io.swagger.annotations.*;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;

import java.util.List;
import javax.validation.constraints.*;
import javax.validation.Valid;
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2019-09-10T18:02:12.946+02:00")

@Api(value = "stat", description = "the stat API")
public interface StatApi {

    @ApiOperation(value = "Statusabfrage", notes = "", response = MonthlyModel.class, tags={ "budget-controller", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "OK", response = MonthlyModel.class),
        @ApiResponse(code = 400, message = "Ungültige Anfrage", response = Void.class),
        @ApiResponse(code = 401, message = "Unauthorized", response = Void.class),
        @ApiResponse(code = 403, message = "Unerlaubter Zugriff", response = Void.class),
        @ApiResponse(code = 404, message = "Not Found", response = Void.class),
        @ApiResponse(code = 500, message = "Ein allgemeiner Serverfehler ist aufgetreten", response = Void.class) })
    
    @RequestMapping(value = "/stat",
        produces = { "*/*", "application/json" }, 
        method = RequestMethod.GET)
    default ResponseEntity<MonthlyModel> askForStatUsingGET( @NotNull@ApiParam(value = "month", required = true) @RequestParam(value = "month", required = true) Integer month, @NotNull@ApiParam(value = "year", required = true) @RequestParam(value = "year", required = true) Integer year) {
        // do some magic!
        return new ResponseEntity<MonthlyModel>(HttpStatus.OK);
    }

}